Build node from scratch rather than copying bitshares image
Originally it was useful to copy the bitshares image, making this
node a quick and simple configuration change. However, since the
change in BitShares to only calculate extra statistics data when
indexed with statistics enabled, it became necessary to do a full
re-index when building the publicnode image.

At this point, it makes more sense to just build from scratch
rather than copying the bitshares image and reindexing its chain.
This is because docker's layered filesystem approach means it's
incredibly wasteful of disk space to start with a fully indexed
chain, delete it, and then make a whole new one, as the layer
containing the entire deleted index is kept as part of publicnode's
stack. Building from scratch means publicnode only has one copy
of the index in its filesystem stack.